--- src/urllib3/packages/__init__.py
@@ -1,3 +1,0 @@
-from . import ssl_match_hostname
-
-__all__ = ("ssl_match_hostname",)
